Added CVAR 'rend-dev-cull-subsectors': Disable non-visible subsector culling (for debug). Made CVARs 'rend-dev-freeze' & 'rend-dev-light-modmatrix' NO_ARCHIVE
Fixed a bug which caused translated textures to flash white after a console RESET
.
Began cleanup of Rend_RenderWallSeg()

Fixed a bug which caused HOM when the viewer moved through a unmasked mid texture that completely filled the floor-ceiling gap on two-sided lines. If the viewer is close enough and the line does not block movement we WONT render a solid seg behind it.

Added the ambient light lower limit to Rend_CalcLightRangeModMatrix.

If viewer is a camera and in the void - don't create occlusion planes for the subsector. Do clear the COLOR_BUFFER_BIT though (we don't want HOM, IF we're in the void)

Improved drawing of debug mobj bounding boxes so they are more transparent the further they get from the viewer (clearer to see).