Further enhanced the self-referencing hack sector detection to cope with a problem illustrated in BLUDWRKS.WAD
Due to the algorithms linear nature, it did not consider self-collision (at which point it would go into an infinite loop).

Changed the algorithm to a non-linear "collection routine" to deal with this shortcoming.

Disabled ALL missing texture replacement (not to be confused with bad texture replacement) due to the fact it was causing Doomsday to crash on some PWADs (TODO: re-implement missing texture replacement using more advanced logic).