Added a new libdeng variable DD_DENG2_CLIENT that determines whether the game is actually a libdeng2 client.
Decided to forgo the full map state serialization for now (too far-reaching
impact to do correctly at this stage); the client loads the map just like the
server does.

The client only spawns the player, no other mobjs.

Once all the platforms are go, this is the end of phase 3.