Began refactoring management of dynlights, dynnodes and dynlists. First steps completed; gamemap_t now takes ownership of them. I also had a bit of clean up.
Note: In the process of which, I found and fixed a subtle (but rare) bug with the brightest-first sorting of dynlist dynnodes (used with non-opaque mid textures) that could result in a segfault. However it is much more likely that result would have been the invalidation of the linked list of dynnodes for dynlist it was attempting to sort (resulting in a memory leak).

The next step; move the dynlists back out of gamemap_t creating a global pool of them (requires moving dynlight_t out of dynnode_t).