Began clean up of rend_main.c and refactoring the concepts I've introduced in the renderer since I joined the project.
Long overdue I know. For that I must apologize. Whilst I was eager to help, it is pretty clear that I lacked the discipline/understanding to implement them cleanly at the time (and probably shouldn't have meddled truth be told).

I'm not going to go overboard with this, as much of it will be redundant once we merge the mapcache with the new-order branch. I plan to do just enough to de-obfuscate and clarify the design. The real refactoring work will be left until later once the fate of much of this code is decided.