Continued clean up of rend_main.c * Implemented further abstractions to the drawing of seg sections so that the algorithm could be split up into discreet stages; determine paramaters, construct geometry, render polygons. This allowed for a few optimizations here and there, resulting in a noticeable, although slight, ~8FPS gain in maps with lots of twosided middle textures. * Changed: Further enhanced the LOS algorithm so that rays may be optionally clipped by opaque middle materials on twosided linedefs. Disabled by default when called by P_CheckSight to retain vanilla DOOM behaviour. * Fixed: Addressed overzealous checking in the oblinked spreading algorithm which prevented spreading over self-referencing linedefs.
Todo: Rework seg division (walldivs) so that divs member of rendseg_t can be discarded. The idea with rendseg_t is that it will ultimately replace seg_t. The difference being, rendseg_t is an abstract object which hides the complexities of determining which properties to use when drawing sections of the sidedef it is bound to.

Combined with rendplane_t (once implemented) they will serve as the division between the new map representation and the renderer; i.e., rendseg_t and rendplane_t are attached to the half-edge DS of Map in libdeng2. So. If we want to retain the existing render list implementation in libdeng, the natural API would exist at RL_AddPoly. The angle clipper would be moved into libdeng2.