Our nodebuilder is now producing a full compliment of real half-edges (only one vertex ptr). That means that where once there were single segs there are now twin-ed half-edges (e.g., single sided linedefs now have half-edges on both sides but only the front will have segs).
In order to get this working I had to make a temporary change to node_t. Currently, the nodebuilder will split the leafs (what were subsectors) into nodes and put the subsector into the right child. The left child is left invalid.

Just as I was about to start resolving this, I began to wonder if it might not be advantageous to leave the configuration as it is; so long as we take care to account for these where necessary, they may prove useful for temporarily attaching geometry in the runtime map editor (e.g., sorting, selection...).

For now, I think I'll sleep on it.