Limit visibility of face_t from the half-edge datastructure to subsystems of the engine that need to know about them. Access to face_t is no longer possible via DMU and has been replaced with DMU_SUBSECTOR.
Todo: The game plugins should be responsible for defining a problem domain (game side) lexicon as an abstraction of solution domain (engine side) objects and interfaces. All these need amount to presently is a set of typedefs and macros in common/include/dmulib.h On the engine side, we should replace all DOOM-derived names for the map data objects as they bare little resemblance to the objects from which their names were taken.