Minor change to de::Lex to allow the single line comment character to be set via the constructor (but with a default of '#' so as not to change existing behaviour).
I intend to use de::Lex in the rewrite of wadconverter::HexenDefParser (ANIMDEFS, MAPINFO etc). If I need to make more changes to the lexical analysis for this syntax; I'll replace the last argument of the de::Lex constructor with an optional Syntax struct (with a default constructor which retains the existing syntax). This way the syntax can be defined for a script type once and then reused.