Implemented mechanics to allow localization of Doomsday Windows Installer
To add a new language/country localization:
* Create a new folder under distrib/win32 named according to the convention:
    Name = <ISO 639-1>-<ISO 3166-1>
* Create a .wxl file in the localized folder for each *.wxs source file
  under distrib/win32 that you wish to localize.
* Edit platform_buildinstaller.py adding a new entry to list of locales
  with the requiste attributes for the new locale.

At present there are only strings for the base-/default- language (en-us)
and the localized versions are currently output to additional .msi files
in distrib/releases. This is not ideal - we don't want to be distributing
multiple installers for the sake of localization.

The next step is to use the localized versions to create a set of transforms
that can then be embedded into The One installer. We would then implement
dynamic language pre-selection by chaining a bootstrapper into the .msi