Embed localized transforms into The One installer msi
Use the localized versions of the Doomsday Windows Installer to generate
a transform which is then embedded into the main .msi Now when executing
the .msi a localized version of the UI can be chosen with the argument:

  TRANSFORMS":<ISO 639-1>-<ISO 3166-1>"

Modified the localization construction so that strings defined for the
en-us locale are used as fallbacks regardless of the current locale.

Make use of the finished default translations for the prebuilt WixUI_Mondo
UI to embed Danish, French and Romanian localizations in the final .msi

Todo: There is presently a codepage mismatch issue with the Chinese and
Slovakian localizations that prevents them from being used.